pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = ""
authors = ["yeezy-na-izi <howone1sky@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_template"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
toml = "^0.10.2"
tortoise-orm = "^0.19.2"
aerich = "^0.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
